% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoDIF.ABWIC.R
\name{plot_lasso_paths}
\alias{plot_lasso_paths}
\title{Plot coefficient paths from LASSO DIF}
\usage{
plot_lasso_paths(
  out,
  nr.lambda = 100,
  highlight = NULL,
  title = "Regularization Paths of DIF Effects",
  ...
)
}
\arguments{
\item{out}{A fitted object returned by \code{lassoDIF()}.}

\item{nr.lambda}{Number of lambda values to evaluate and display (default is 100).}

\item{highlight}{Optional: indices of items to highlight in color.}

\item{title}{Main title of the plot.}

\item{...}{Additional graphical parameters passed to \code{matplot()}.}
}
\value{
A base R plot of coefficient paths.
}
\description{
This function displays coefficient trajectories from LASSO-regularized DIF detection.
}
